Remove unnecessary indentation.
In Markdown, code blocks can be done either by indenting a block by 4
spaces (or a tab), or by using a fenced code block.

When using fenced code block, it's possible to specify the language for
highlighting, and it's not necessary to indent the code. That
indentation is preserved as part of the content, which results in odd
formatting here since it's not intended.

See
https://help.github.com/articles/creating-and-highlighting-code-blocks/
for details on fenced code blocks.

Also fix minor formatting inconsistencies.